third_party/WebKit/LayoutTests/http/tests/inspector/network/load-resource-for-frontend.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    var loggedHeaders = {
        "cache-control": true,
        "expires": true,
        "last-modified": true,
        "pragma": true
    };
    function testLoadForURL(url, headers, next)
    {
        InspectorTest.recordNetwork();
        InspectorTest.addResult("Loading resource from " + url);
        NetworkAgent.loadResourceForFrontend(WebInspector.resourceTreeModel.mainFrame.id, url, headers, callback);

        function callback(error, statusCode, headers, content)
        {
            if (error)
                InspectorTest.addResult("Failed: " + error);
            else {
                InspectorTest.addResult("Success: " + statusCode);
                var headersArray = [];
                for (var name in headers) {
                    var nameLower = name.toLowerCase();
                    if (nameLower in loggedHeaders)
                        headersArray.push(nameLower);
                }
                headersArray.sort();
                InspectorTest.addResult("Logged headers: " + headersArray.join(", "));
                InspectorTest.addResult("Content: \"" + content + "\"");
            }
            InspectorTest.runAfterPendingDispatches(dumpNetworkRequest);

            function dumpNetworkRequest()
            {
                WebInspector.inspectorView._showPanel("network");
                var request = InspectorTest.networkRequests().pop();
                InspectorTest.addResult("Dumping request: " + request.url);
                InspectorTest.addResult("  statusCode: " + request.statusCode);
                next();
            }
        }
    }

    InspectorTest.runTestSuite([
        function testNetworkResourceNonCached(next)
        {
            testLoadForURL("http://localhost:8080/inspector/network/resources/resource.php", undefined, next);
        },

        function testNetworkResourceCached(next)
        {
            testLoadForURL("http://localhost:8080/inspector/network/resources/resource.php?cached=1", undefined, next);
        },

        function testNetworkResourceIfModifiedSince(next)
        {
            testLoadForURL("http://localhost:8080/inspector/network/resources/resource.php?cached=1", { "if-modified-since": new Date().toUTCString() }, next);
        },

        function testResourceFromIncorrectURL(next)
        {
            testLoadForURL("resource.php", undefined, next);
        },

        function testResourceFromFileScheme(next)
        {
            testLoadForURL("file:///var/www/index.html", undefined, next);
        },

        function testResourceWithCookie(next)
        {
            InspectorTest.makeSimpleXHR("GET", "http://127.0.0.1:8000/inspector/network/resources/set-cookie.php", false, callback)

            function callback()
            {
                testLoadForURL("http://127.0.0.1:8000/inspector/network/resources/print-cookie.php", undefined, next);
            }
        },
    ]);
}
</script>
</head>
<body onload="runTest()">
</body>
</html>
