third_party/WebKit/LayoutTests/http/tests/inspector/network/load-resource-for-frontend-expected.txt


Running: testNetworkResourceNonCached
Loading resource from http://localhost:8080/inspector/network/resources/resource.php
Success: 200
Logged headers: cache-control, expires, pragma
Content: "Hello world"

Running: testNetworkResourceCached
Loading resource from http://localhost:8080/inspector/network/resources/resource.php?cached=1
Success: 200
Logged headers: cache-control, expires, last-modified
Content: "Hello world"

Running: testNetworkResourceIfModifiedSince
Loading resource from http://localhost:8080/inspector/network/resources/resource.php?cached=1
Success: 304
Logged headers: 
Content: ""

Running: testResourceFromIncorrectURL
Loading resource from resource.php
Failed: Couldn't create a loader

Running: testResourceFromFileScheme
Loading resource from file:///var/www/index.html
Failed: Can not load local file
